+++ fbw-a32nx/src/base/flybywire-aircraft-a320-neo/html_ui/Pages/VCockpit/Instruments/Airliners/FlyByWire_A320_Neo/CDU/A320_Neo_CDU_InitPage.js
@@ -413,6 +413,7 @@ class CDUInitPage {
 
         };
 
+        let fuelPageTitle = new Column(10, "INIT");
         const fuelPlanTopTitle = new Column(23, "", Column.amber, Column.right);
         const fuelPlanBottomTitle = new Column(23, "", Column.amber, Column.right);
         if (mcdu._zeroFuelWeightZFWCGEntered && !mcdu._blockFuelEntered) {
@@ -426,6 +427,7 @@ class CDUInitPage {
             };
         }
         if (mcdu._fuelPlanningPhase === mcdu._fuelPlanningPhases.IN_PROGRESS) {
+            fuelPageTitle = new Column(1, "INIT FUEL PLANNING");
             fuelPlanTopTitle.update("BLOCK ", Column.green);
             fuelPlanBottomTitle.update("CONFIRM", Column.green);
             mcdu.onRightInput[2] = async () => {
@@ -532,6 +534,7 @@ class CDUInitPage {
         }
 
         if (CDUInitPage.fuelPredConditionsMet(mcdu)) {
+            fuelPageTitle = new Column(1, "INIT FUEL PREDICTION");
             fuelPlanTopTitle.text = "";
             fuelPlanBottomTitle.text = "";
 
@@ -693,7 +696,7 @@ class CDUInitPage {
 
         mcdu.setTemplate(FormatTemplate([
             [
-                new Column(5, "INIT FUEL PRED")
+                fuelPageTitle
             ],
             [
                 new Column(0, "TAXI"),